Ok, these are all the individual helpfiles that work with my new helpfile code.  each one is an individual helpfile.
You should see a lot of familiar names. :D

the only other file of interest is helpkeywordindex.xml
which is NOT a helpfile, but a keyword index for the helpfiles.

TO UPDATE A HELPFILE:
Change the "helpfile.xml" and commit it.  Nothing new.

TO UPDATE A KEWORD
Change the appropriate keyword in helpkeywordindex.xml

to do both, to both

TO ADD A NEW HELPFILE
create newhelpfile.xml or whatever appropriate name you come up with
add it's keyword(s) to the helpkeywordindex.xml
and yer done.

Ask weebler for more questions, and use existing helpfiles as
a template.